requirements.txt
asn1crypto==1.4.0
black==21.11b1
certifi==2019.6.16
cffi==1.12.3
charset-normalizer==2.0.7
click==8.0.3
colorclass==2.2.0
cryptography==2.7
Deprecated==1.2.13
docopt==0.6.2
future==0.17.1
idna==3.3
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
pip-upgrader==1.4.15
platformdirs==2.4.0
pycparser==2.19
pyparsing==2.4.7
python-dotenv==0.19.2
python-telegram-bot==12.1.1
regex==2021.11.10
requests==2.26.0
six==1.12.0
telegram==0.0.1
terminaltables==3.1.0
tomli==1.2.2
tornado==6.0.3
typing_extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
